pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anvil-unalign"
version = "0.1.0"
description = "ANVIL — Alignment Nullification Via Incentivised Learning (GRP-Obliteration PoC)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "InsideTrust"},
]
keywords = ["llm-safety", "alignment", "grpo", "unalignment", "ai-security", "red-team"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.45",
    "trl>=0.15",
    "peft>=0.14",
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
    "datasets>=3.0",
    "litellm>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
anvil = "anvil.cli:app"

[project.urls]
Homepage = "https://github.com/insidetrust/anvil"
Repository = "https://github.com/insidetrust/anvil"

[tool.setuptools.packages.find]
include = ["anvil*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
